+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -106,7 +106,7 @@ public class KNNConstants {
     public static final double MAXIMUM_CONFIDENCE_INTERVAL = 1.0;
     public static final String LUCENE_SQ_BITS = "bits";
     public static final int LUCENE_SQ_DEFAULT_BITS = 7;
-    public static final int OPTIMIZED_SCALAR_QUANTIZER_DEFAULT_BITS = 1;
+    public static final int LUCENE_SCALAR_QUANTIZER_DEFAULT_BITS_AFTER_V360 = 1;
     public static final String MAX_CONNECTIONS = "max_connections";
     public static final String BEAM_WIDTH = "beam_width";