+++ src/main/java/org/opensearch/knn/index/mapper/CompressionLevel.java
@@ -151,11 +151,11 @@ public enum CompressionLevel {
             return RescoreContext.builder().oversampleFactor(FLAT_OVERSAMPLE_FACTOR).userProvided(false).build();
         }
         if (modesForRescore.contains(mode)) {
-            // Special handling for Optimized Scalar Quantizer (x32 compression)
-            // Engine check is temporary until optimized scalar quantizer is finalized for FAISS as well
+            // Special handling for Lucene Scalar Quantizer (x32 compression)
+            // Engine check is temporary until binary scalar quantizer is finalized for FAISS as well
             if (this == x32 && engine == KNNEngine.LUCENE && version.onOrAfter(Version.V_3_6_0)) {
                 return RescoreContext.builder()
-                    .oversampleFactor(RescoreContext.OVERSAMPLE_FACTOR_DEFAULT_FOR_OPTIMIZED_SCALAR_QUANTIZER)
+                    .oversampleFactor(RescoreContext.OVERSAMPLE_FACTOR_DEFAULT_FOR_LUCENE_SCALAR_QUANTIZER_AFTER_V360)
                     .userProvided(false)
                     .build();
             }